+++ CHANGELOG.md
@@ -41,6 +41,6 @@
   
 ## March 4  2018 0.5.0
     * Updated criteria to support comprehensive conditional expression.
-    * Reorganized service and dependencies (major refactoring)
+    * Reorganized services and dependencies
     * Minor fixes